requirements.txt
streamlit
streamlit-lottie
requests
plotly
wordcloud
folium
streamlit_folium
openpyxl
scikit-learn
python-docx
python-dotenv